Rearrange PortList and PortRange test to new structure
