+++ tests/protocols/gmpv208/entities/port_lists/__init__.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2021 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from .test_clone_port_list import GmpClonePortListTestMixin
+from .test_create_port_list import GmpCreatePortListTestMixin
+from .test_create_port_range import GmpCreatePortRangeTestMixin
+from .test_delete_port_list import GmpDeletePortListTestMixin
+from .test_delete_port_range import GmpDeletePortRangeTestMixin
+from .test_get_port_list import GmpGetPortListTestMixin
+from .test_get_port_lists import GmpGetPortListsTestMixin
+from .test_modify_port_list import GmpModifyPortListTestMixin
