+++ tests/protocols/gmpv208/entities/port_lists/test_clone_port_list.py
@@ -16,20 +16,20 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import GvmError
+from gvm.errors import RequiredArgument
 
 
-class GmpDeletePortRangeTestCase:
-    def test_delete(self):
-        self.gmp.delete_port_range('a1')
+class GmpClonePortListTestMixin:
+    def test_clone(self):
+        self.gmp.clone_port_list('a1')
 
         self.connection.send.has_been_called_with(
-            '<delete_port_range port_range_id="a1"/>'
+            '<create_port_list>' '<copy>a1</copy>' '</create_port_list>'
         )
 
     def test_missing_id(self):
-        with self.assertRaises(GvmError):
-            self.gmp.delete_port_range(None)
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_port_list('')
 
-        with self.assertRaises(GvmError):
-            self.gmp.delete_port_range('')
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_port_list(None)
